perf(interpreter):Improve ATC interpreter's cache efficiency
KAG-5152
