+++ benches/build.rs
@@ -12,6 +12,10 @@ use uuid::Uuid;
 
 const N: usize = 1000;
 
+fn make_uuid(a: usize) -> String {
+    format!("8cb2a7d0-c775-4ed9-989f-{:012}", a)
+}
+
 fn criterion_benchmark(c: &mut Criterion) {
     let mut schema = Schema::default();
     schema.add_field("a", Type::Int);
@@ -24,7 +28,9 @@ fn criterion_benchmark(c: &mut Criterion) {
             let mut router = Router::new(&schema);
             for i in 0..N {
                 let expr = format!("((a > 0 || a < {}) && a != 0) && a == 1", N + 1);
-                router.add_matcher(N - i, Uuid::new_v4(), &expr).unwrap();
+                let variant = make_uuid(i);
+                let uuid = Uuid::try_from(variant.as_str()).unwrap();
+                router.add_matcher(N - i, uuid, &expr).unwrap();
             }
             router
         });
